requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain_openai
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
python-dotenv
typesense
langgraph
Pillow
open-clip-torch
docx2txt